removing firstAvailableDate/customerAcceptanceDate check for GWeekly
This check is not necessary as the current calculation will not
let you create a holiday stop before fulfilment starts.

This check also causes problems as the customerAcceptance date is the
start of the full subscription not the start of 6-for-6 subscription
hence holiday stops would not be allowed in the 6-for-6 period.