fixing holiday stop cancellation endpoints handling of processed stops
if a holiday stop is processed but the bill that the refund
should be applied to is on or after the cancellation date.

The bill will not be generated and the user will not receive the refund.

This change will include those holiday stops in the holiday stop api
cancellation endpoints so the user will get a manual refund for those
stops