fixing credit invoice date calculation
there are two ways of calculating the start date of a charge, either
using effective start date or the driving it manually from a combination
of other fields which way works seems to be random.

This change selects the method that correctly predicts the
processedThrough date which is either the last invoice date for that
charge or it is the day after the last day of the charge if the
charge has ended.