Fix Scala Steward 'gitignore' error
The Guardian's Scala Steward GitHub Action job has been erroring for the
past 5 days (since 9:55am GMT on Thursday 19th January 2023 GMT) -
looking in the logs (https://github.com/guardian/scala-steward-public-repos/actions/runs/3957277297/jobs/6777445592)
we can see that Scala Steward is erroring on the
https://github.com/guardian/support-service-lambdas repo:

```
2023-01-19 10:02:07,438 ERROR Steward guardian/support-service-lambdas failed
  java.io.IOException: 'GIT_ASKPASS=/home/runner/scala-steward/askpass.sh "SBT_OPTS=-Xmx2048m -Xss8m -XX:MaxMetaspaceSize=512m" git -c core.hooksPath=/dev/null add /home/runner/scala-steward/workspace/repos/guardian/support-service-lambdas/.git-blame-ignore-revs' exited with code 1
  The following paths are ignored by one of your .gitignore files:
  .git-blame-ignore-revs
  hint: Use -f if you really want to add them.
  hint: Turn this message off by running
  hint: "git config advice.addIgnoredFile false"
```

The error message tells us that Scala Steward is trying to add a file
called `.git-blame-ignore-revs` to the guardian/support-service-lambdas
repo, but the `git` command is rejecting this, because back in October
2020, https://github.com/guardian/support-service-lambdas/pull/734
added `.gitignore` configuration that rejects all files starting with
a `.` (hidden files, in UNIX - https://en.wikipedia.org/wiki/Hidden_file_and_hidden_directory ).

Why is Scala Steward trying to add a `.git-blame-ignore-revs` file
to the repo? This was a feature added to Scala Steward in July 2022
with https://github.com/scala-steward-org/scala-steward/pull/2652,
fixing https://github.com/scala-steward-org/scala-steward/issues/2567:
now when Scala Steward creates 'reformatting' commits with scalafmt, it
adds the corresponding commit id to `.git-blame-ignore-revs`, which is
a file respected by both `git` & GitHub:

* https://michaelheap.com/git-ignore-rev/
* https://docs.github.com/en/repositories/working-with-files/using-files/viewing-a-file#ignore-commits-in-the-blame-view

Now, when you try to `blame` a file in `git` or GitHub, the results will
*leave out* reformatting commits, so long as they're listed in
`.git-blame-ignore-revs`, which is nice - so it's not a bad thing that
Scala Steward is trying to create this file. However, we need to change
this repo's `.gitignore` file to allow Scala Steward to create the file.

The configuration added in //github.com/guardian/support-service-lambdas/pull/734
was 'reject hidden files by default' - however, it now turns out that of
the hidden files used in this project, _more_ of them are desired committed
to the repo than not! So, it's better to switch the config to 'accept
hidden files by default' - and just specify the 2 hidden folders we don't want
to commit.