Restructure alarm mappings (#2239)
Instead of mapping apps to teams, invert the data structure to be an object
with teams as keys and arrays of apps as values. I think this is a nice data
structure to work with when adding apps. Then we transform this back to the
original format to do the lookup from app to team.