Check TypeScript formatting in CI (#2359)
Check code formatting as part of the build.

This also includes a change to the TS CI workflow configuration to run tests,
check formatting etc on common code to a separate build job, so that we don't
repeat for every sub-project. Without doing this, a failed test or formatting
issue will cause all sub-project builds to fail which I think is confusing.