Enforce PNPM with corepack (#2444)
Just like with Node, we want developers to be using a consistent package
manager/version, to ensure the same experience on our laptops and also in CI.
In #2443 a couple of non-pnpm lockfiles were removed. In this PR we start
enforcing usage of pnpm so this shouldn't happen again. Now if I try to run
yarn for example I get an error:

Usage Error: This project is configured to use pnpm

Also use corepack in CI (and specify for local development in the README) to
manage package managers.