Ensure that 4xx response bubble up correctly from identity (#2595)
Previously API gateway wasn't happy with the "type" property which was
added to the response resulting in a 502. This only existed so that
internally we could tell what kind of response we'd got back. Now the
response for a failed auth request is nested, so the type doesn't appear
in the response we give back to API gateway.