Move top level modules into utils
Supporting files at the top level of the modules dir was a bit of an
exception and meant we needed an additional rule in the jest config.
It's only one line but it doubled the cognitive load of understanding
how the module mappings worked, and it lead to some inconsistency of how
modules were imported.

Because of the exception, it also meant top level modules tests weren't
getting run in CI.