Add a new endpoint `GET /benefits/<identity id>` to the user benefits API (#2606)
Add a new endpoint /benefits/<identity id> which is authed using an x-api-key
header (enforced in API Gateway). Because the auth is performed differently, at
the API Gateway, I've migrated to using the GuApiGatewayWithLambdaByPath with a
lambda per endpoint.

This will be called from support-frontend to prevent a user who already has a
the adFree benefit from purchasing Guardian Ad-Lite.