Type active and inactive products in the catalog distinctly (#2628)
In the product catalog a distinction is already made between active and
inactive products, where active == products we're still selling and inactive ==
products we aren't, but which people may still have. However, this distinction
isn't available at the type level.

In support-frontend, where we consume the product catalog data, we use the
ProductKey type. Currently this includes all products and in cases where we
have a Record<ProductKey, A> we have to provide an entry for all products, even
inactive ones, which doesn't really make sense if we aren't selling them as the
data won't be used. We could move to a Partial<Record<ProductKey, A>> but this
means we have to handle undefined when looking up by a product (even if we know
in practice this won't happen. I think a better approach would be to have
separate types for active and inactive products. After this change we can
support Record<ActiveProductKey, A> meaning we'd only have to provide data
about active products and not deprecated products.

So, this PR splits the single generated typeObject to two - one for active
products and one for inactive. We then derived separate (and combined) types
from these two objects.

This PR also adds GuardianAdLite, which is now in the type object (an active
project) since it's now in the Zuora data.