Metric Push API: Only alarm on 2xx errors (#2655)
I'm attempting to make the metric push API alarm less noisy. The strategy will
be to return a non-2xx response code for clients we don't care about. In order
to make this work we'll need to be able to alarm only on genuine 2xx responses.
Even without this requirement, I think it makes sense to only alarm on 2xxs. If
for whatever reason a client makes requests to this endpoint which we're not
set up to handle (incorrect HTTP method for example) I don't think that should
trip the alarm.