Migrate metric-push-api cloudformation to cdk/ci-typescript workflow (#2664)
This creates a cdk stack which simply wraps the existing cloudformation
as described here:
https://github.com/guardian/cdk/blob/main/docs/migration-guide.md

Move metric-push-api over to the ci-typescript workflow (when I add the lambda
we'll have to use this).

The reason we create a dummy package script for metric-push-api is because the
ci-typescript workflow expects all projects to have a cloudformation and lambda
artefact. There's no lambda part to metric-push-api (yet, I'm planning on
adding one in a future PR) so doing this just to keep the workflow happy so
there's a file where it expects, without having to introduce an exception.

We supply the cloudFormationStackName param to the (cdk) stack. Akash
recommended this so that gu cdk knows what to diff against.