Extract cloudwatch from ticket-tailor-webhook to module (#2684)
Extract cloudwatch push metric code from ticket-tailor-webhook to a module. I
want to use this in another workspace.

Note: a few import paths have changed as I wanted to be able to import
@modules/aws/cloudwatch without src in the path. I updated the jest module
mappings for ticket-tailor-webhook and this had a knock on effect of having to
update the secrets-manager paths too (also to remove the src). I think this
leaves the code in a (marginally) nicer state!