Metric Push API Lambda (#2683)
This PR introduces a lambda to handle metric-push-api requests. We now check
the referer in the lambda and if it matches a pre-defined list (the support
site) then emit a custom metric. Alarm on this metric. This gives us much more
control over filtering requests which we think are legit and actually want to
emit the metric for.