Don't run cdk package for every TypeScript lambda
I noticed that the cdk package step is by far the slowest part of each
TypeScript lambda build. The package script does type checking,
formatting checking, linting, runs the tests and synths the
Cloudformation. We don't need to do all of this for every lambda, so
instead run all the checks in the common job and only synth the
Cloudformation in the invidual lambda jobs (which we need to upload to
Riff Raff.