Revert "chore(deps): bump @aws-sdk/client-ssm from 3.777.0 to 3.787.0" (#2806)
Revert #2790 as it's causing a type error which is breaking tests on currently
open PRs. Previously this wasn't a blocker but we now require all build checks
to pass (as we should!)

I think the source of the error is to do with the fact that we have multiple
different @aws-sdk libs in the project at different versions. This has caused
multiple versions of @types/smithy to exist in the project, according to
https://www.npmjs.com/package/aws-sdk-client-mock#caveats this can cause type
errors like the one we're seeing:

error TS2345: Argument of type 'typeof SecretsManagerClient' is not assignable
to parameter of type 'InstanceOrClassType<Client<ServiceInputTypes,
MetadataBearer, SmithyResolvedConfiguration<HttpHandlerOptions>>>'