Two distinct payment methods for PPCP
The first step in the rollout of the PayPal/Zuora migration will be to
use the new PayPal Complete Payments gateway with the existing SDK and
API, passing a BAID (just like with PayPal Express).

Then we'll gradually migrate to passing a vault payment token to the
gateway, with the token generated by using the new SDK and REST API.
Initially this payment method won't be used, but it makes the migration
path clear.