Add new PayPal Complete Payments payment gateway & payment methods (#3150)
This PR adds a new PayPal Complete Payments Zuora payment gateway (PayPalCP)
and two distinct payment methods which use this gateway:

1. The first step in the rollout of the PayPal/Zuora migration will be to use
the new PayPal Complete Payments gateway with the existing SDK and API, passing
a BAID (just like with PayPal Express). This is the first new payment method.

2. Then we'll gradually migrate to passing a vault payment token to the
gateway, with the token generated by using the new SDK and REST API. In this
case the payment gateway used is the same, but slightly different parameters
are passed. This is the second new payment method.

Initially 2 won't be used, but it makes the migration path clear to add them
both now.

This PR in and of itself doesn't change anything - nothing will use the new
payment methods (and therefore the new gateway) until changes in
support-frontend are made.